+++ SofaGuidewireNav/simulation/SimServer.py
@@ -10,9 +10,9 @@ import time
 import pickle
 import numpy as np
 
-# <SofaGuidewireNav>/Package/simulation/../../
+# <SofaGuidewireNav>/SofaGuidewireNav/simulation/../../
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__))+"/../../")
-from Package.utils import mkdir, root_dir
+from SofaGuidewireNav.utils import mkdir, root_dir
 
 
 class Server():
