+++ SofaGuidewireNav/simulation/scene.py
