--- __init__.py
@@ -1,4 +1,0 @@
-import SofaGuidewireNav.Package.utils as utils
-from SofaGuidewireNav.Package.simulation.SimServer import SimController
-
-__all__ = ["utils", 'SimController']
