+++ example_module.py
@@ -1,5 +1,5 @@
 from SofaGuidewireNav import SimController
-from SofaGuidewireNav import utils
+from SofaGuidewireNav.utils import SaveImage
 
 sim = SimController(timeout=10)
 sim.open()
@@ -9,5 +9,5 @@ for i in range(500):
     sim.action(translation=1, rotation=0.1)
     sim.step(realtime=False)
     image = sim.GetImage()
-    utils.SaveImage(image=image, filename=f'image/image_{i}.jpg')
+    SaveImage(image=image, filename=f'image/image_{i}.jpg')
 sim.close()
